+++ bauh/gems/appimage/controller.py
@@ -223,7 +223,18 @@ class AppImageManager(SoftwareManager):
                             for tup in cursor.fetchall():
                                 for app in res.installed:
                                     if app.name.lower() == tup[0].lower() and (not app.github or app.github.lower() == tup[1].lower()):
-                                        app.update = LooseVersion(tup[2]) > LooseVersion(app.version) if tup[2] else False
+                                        continuous_version = app.version == 'continuous'
+                                        continuous_update = tup[2] == 'continuous'
+                                        if continuous_version and not continuous_update:
+                                            app.update = True
+                                        elif continuous_update and not continuous_version:
+                                            app.update = False
+                                        else:
+                                            try:
+                                                app.update = LooseVersion(tup[2]) > LooseVersion(app.version) if tup[2] else False
+                                            except:
+                                                app.update = False
+                                                traceback.print_exc()
 
                                         if app.update:
                                             app.latest_version = tup[2]